requirements.txt
pandas>=2.2.3
numpy>=1.26.4
matplotlib>=3.8.4
seaborn>=0.13.2
google-generativeai>=0.3.2
sqlalchemy>=2.0.30
chromadb>=0.4.24
python-dotenv>=1.0.1
sentence-transformers>=2.6.1
plotly>=5.22.0
pytest>=7.4.3
chardet>=5.2.0
openpyxl>=3.1.2
xlrd>=2.0.1
pymysql>=1.1.0
psycopg2-binary>=2.9.9
